+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [v0.3.0](https://github.com/pnezis/tucan/tree/v0.3.0) (2024-01-03)
 
 ### Added